Important information to prevent frustration
It took me a whole week to get this server running behind a reverse proxy. There's lots of stack exchange threads of people having problems getting socket.io servers working in combination with reverse proxying and very few of these threads lead to a solution.

I provided some stack exchange solutions to the problems I came across on this journey, but I think it would also be really helpful to have the given hin in this server example script.